py-175/file-based-cms/pyproject.toml
[project]
name = "file-based-cms"
version = "0.1.0"
description = ""
authors = [
    {name = "Karishma Tank",email = "tankkarishma@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "markdown (>=3.10,<4.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
